requirements.txt
Flask
PyPDF2 < 3.0
camelot-py
fastapi
ghostscript
loguru
loko-client
loko-extensions
pymupdf
python-multipart
table_ocr
